refactor(Validator.js): make class to file
