+++ src/App.js
@@ -1,6 +1,7 @@
-import { Console, Random } from "@woowacourse/mission-utils";
+import { Console } from "@woowacourse/mission-utils";
 import RACING_VARIABLES from './constants/RacingVariables';
 import RaceController from "./controllers/RaceController";
+import Validator from "./utils/Validator";
 
 
 
@@ -28,24 +29,6 @@ class UserInput{
     }
   }
 }
-
-class Validator {
-
-  validateNameLength(names) {
-    names.forEach(name => {
-      if (name.length > 5) {
-        throw new Error(`[ERROR]: 이름 "${name}"이 5자를 초과했습니다.`);
-      }
-    });
-  }
-
-  validateRound(count){
-    if (count > RACING_VARIABLES.MAXIMUM_ROUND){
-      throw new Error(`[ERROR]: 최대 라운드 횟수 "${RACING_VARIABLES.MAXIMUM_ROUND}"회를 초과했습니다.`);
-    }
-  }
-}
-
 
 class DisplayController{
